LayoutTests/fast/forms/label/labels-custom-property.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description('Test that we can set and get custom properties on the labels attribute. ');

var parentDiv = document.createElement('div');

parentDiv.innerHTML = '<button id="btn"></button><label id="lbl1" for="btn"></label><label id="lbl2" for="btn"></label>';

document.body.appendChild(parentDiv);

labels = document.getElementById("btn").labels;
labels.customProperty = 1;
gc();
shouldBe('labels.customProperty', '1');
shouldBe('document.getElementById("btn").labels.customProperty', '1');
</script>
<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
