third_party/blink/web_tests/fast/forms/label/labels-custom-property-expected.txt

Test that we can set and get custom properties on the labels attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS labels.customProperty is 1
PASS document.getElementById("btn").labels.customProperty is 1
PASS successfullyParsed is true

TEST COMPLETE
